начал перенос логики в PointController, перенёс удаление поинта
